+++ integration-tests/src/test.rs
@@ -12,8 +12,8 @@ use indoc::indoc;
 use libfxrecord::error::ErrorMessage;
 use libfxrecord::net::*;
 use libfxrecorder::proto::{RecorderProto, RecorderProtoError};
+use libfxrunner::osapi::ShutdownProvider;
 use libfxrunner::proto::{RunnerProto, RunnerProtoError};
-use libfxrunner::shutdown::ShutdownProvider;
 use libfxrunner::taskcluster::{Taskcluster, TaskclusterError};
 use reqwest::StatusCode;
 use serde_json::Value;
